Tests files were moved to folders
