+++ client/src/components/Chat/__test__/Webcam.test.js
@@ -1,4 +1,4 @@
-import Webcam from './Webcam';
+import Webcam from '../Webcam';
 import React from "react";
 import ReactDOM from "react-dom";
 import { screen } from '@testing-library/react';
